+++ libs/pavex_runtime/src/request.rs
@@ -1,7 +1,12 @@
+//! An incoming HTTP request.
 use http::{HeaderMap, Method, Uri, Version};
 
 #[non_exhaustive]
 #[derive(Debug)]
+/// All the information that is transmitted as part of an HTTP request ahead of the body.
+///
+/// It includes the [method](Method), the [URI](Uri), 
+/// the [HTTP version](Version), and the [headers](HeaderMap).
 pub struct RequestHead {
     pub method: Method,
     pub uri: Uri,
