+++ libs/pavex_runtime/src/response/body/errors.rs
@@ -1,0 +1,2 @@
+//! Errors that can arise when building a [`Response`](crate::response::Response) body.
+pub use super::json::JsonSerializationError;
